provider/MetricsProvider2 : refactoring
