+++ src/utils/MetricsProvider.jsx
@@ -1,13 +1,15 @@
 import { createContext, useContext, useEffect, useState } from "react";
 import defaultTags from "../data/standardTags.json";
-import { ToggleGroup, ToggleGroupItem } from "@/components/ui/toggle-group";
 import { useUserContext } from "./contexts/UserProvider";
-// import { Button } from "@/components/ui/button";
-// import Delete from "/src/assets/icons/delete-2-svgrepo-com.svg";
+
+//-----------
+import { getAllCheckins } from "./services/getAllCheckins";
+import { getStatisticsByFamily } from "./services/getStatisticsByFamily";
+import { getStatisticsByTag } from "./services/getStatisticsByTag";
+import { renderTagListbyCategoryMetrics } from "./helpers/renderTagListbyCategoryMetrics";
+import { handleTagToggle } from "./helpers/handleTagToggle";
 
 const MetricsContext = createContext();
-const baseURL = import.meta.env.VITE_baseURL;
-const basePathThree = import.meta.env.VITE_basePathThree;
 
 const MetricsProvider = ({ children }) => {
   // ----------------------------------- States---------------------------------
@@ -17,8 +19,6 @@ const MetricsProvider = ({ children }) => {
   // statisticTwo
   const [metricsTwoStatus, setMetricsTwoStatus] = useState(1);
   // statisticTwo
-
-  const [disableTag, setDisableTag] = useState(false);
 
   // selected feelingsFamily - statisticOne
   const [selectedFeelingsFamily, setSelectedFeelingsFamily] = useState(null);
@@ -47,8 +47,7 @@ const MetricsProvider = ({ children }) => {
 
   // -------------------------- States from UserProvider --------------------------
 
-  const { customTags, setCustomTags, handleDeactivateCustom } =
-    useUserContext();
+  const { customTags } = useUserContext();
 
   // ----------------------------------- Values---------------------------------
   // statisticOne
@@ -58,63 +57,35 @@ const MetricsProvider = ({ children }) => {
   const maxMetricsTwoStatus = 2;
 
   // ----------------------------------- Function - fetch getAllCheckins---------------------------------
-
   // get total of check ins
-  const getAllCheckins = async () => {
+  const loadCheckinsTotal = async () => {
     try {
-      const res = await fetch(`${baseURL}${basePathThree}checkins`, {
-        credentials: "include",
-      });
-
-      if (!res.ok) throw new Error("Error fetching check-ins");
-      const data = await res.json();
-      setCheckIn(data.data.length);
+      await getAllCheckins(setCheckIn);
     } catch (error) {
-      console.error(error);
+      console.error("Fehler beim Abrufen der Anzahl der check-ins:", error);
     }
   };
 
   // ----------------------------------- Function -  fetch getStatisticsByFamily ---------------------------------
 
   // get statistics by emotion family
-  const getStatisticsByFamily = async () => {
+  const loadStatisticsByFamily = async () => {
     try {
-      const res = await fetch(
-        `${baseURL}${basePathThree}stats/family/?family=${selectedFeelingsFamily}`,
-        {
-          credentials: "include",
-        }
+      await getStatisticsByFamily(
+        setStatisticsByFamily,
+        selectedFeelingsFamily
       );
-      if (!res.ok) throw new Error("Error fetching statistics by family");
-      const data = await res.json();
-      setStatisticsByFamily(data.data);
     } catch (error) {
-      console.error(error);
+      console.error("Fehler beim Abrufen der Statistiken:", error);
     }
   };
 
   // ----------------------------------- Function -  fetch getStatisticsByTag ---------------------------------
-  const getStatisticsByTag = async (tagName) => {
-    const baseURL = import.meta.env.VITE_baseURL;
-    const path = import.meta.env.VITE_basePathThree;
+  const loadStatisticsByTag = async (tagName) => {
     try {
-      const res = await fetch(`${baseURL}${path}stats/tag?tag=${tagName}`, {
-        credentials: "include",
-      });
-      const data = await res.json();
-      setStatisticsByTag(data.data);
-      return data;
+      await getStatisticsByTag(tagName, setStatisticsByTag);
     } catch (error) {
-      console.error(error);
-    }
-  };
-
-  // Fetch function for fetching statistics by tag
-  const fetchStatsByTag = async (tagName) => {
-    const data = await getStatisticsByTag(tagName);
-    if (data && data.data && data.data.stats) {
-      console.log("data.data.stats:", data.data.stats); // debugging
-      // TODO: transform the data to the format needed for the statistic
+      console.error("Fehler beim Abrufen der Statistiken:", error);
     }
   };
 
@@ -140,7 +111,7 @@ const MetricsProvider = ({ children }) => {
     if (Number(checkIn) >= 7) {
       setShowMetricsOne(true);
       // fetch is not necessary if there are less than 7 check-ins
-      if (selectedFeelingsFamily) getStatisticsByFamily();
+      if (selectedFeelingsFamily) loadStatisticsByFamily();
     }
   }, [selectedFeelingsFamily, checkIn]);
 
@@ -149,7 +120,7 @@ const MetricsProvider = ({ children }) => {
     if (Number(checkIn) >= 7) {
       setShowMetricsTwo(true);
       // fetch is not necessary if there are less than 7 check-ins
-      if (selectedTag) getStatisticsByTag();
+      if (selectedTag) loadStatisticsByTag();
     }
   }, [selectedTag, checkIn]);
 
@@ -160,84 +131,30 @@ const MetricsProvider = ({ children }) => {
         setMetricsTwoStatus((prevStat) => prevStat + 1);
       }
     } catch (error) {
-      console.log("error: ", error);
+      console.log("error loading statistics2: ", error);
     }
   };
 
   // ----------------------------------- Function - handle selcting/deselecting tags ---------------------------------
-  const handleTagToggle = (tag) => {
-    setSelectTag((prevSelectedTag) => {
-      // prevSelectedTag is the previous state of the selectedTag array
-
-      const isSelected = prevSelectedTag?.name === tag.name; // check if the tag is already selected (it has been clicked before)...
-
-      if (isSelected) {
-        // if the tag is selected again it is deselected
-        return null;
-      } else {
-        handleLoadStatisitcsTwo();
-        return tag;
-      }
-    });
-  };
-  // console.log({ selectedTag });
+  const selectTag = (tag) =>
+    handleTagToggle(tag, setSelectTag, handleLoadStatisitcsTwo);
 
   // ----------------------------------- Function - render tags ---------------------------------
 
   // The renderTags function takes a category as a parameter (e.g. “when”, “where”) and searches the standardTags array to find the matching object with the desired category. As soon as the matching object has been found, it renders the singleStandardTags as li elements
-  const renderTagListbyCategoryMetrics = (category) => {
-    const categoryTags = standardTags.find((tag) => tag.category === category);
-    const categoryTagsCustom = customTags[category];
-
-    return categoryTags || categoryTagsCustom ? (
-      <div>
-        <ToggleGroup type="multiple" className="flex flex-wrap gap-3">
-          {categoryTags.singleStandardTags.map((tag, index) => (
-            <ToggleGroupItem
-              key={index}
-              value={tag}
-              aria-label={tag}
-              className={`cursor-pointer p-2 border`}
-              data-state={selectedTag?.name === tag ? "on" : ""}
-              onClick={() =>
-                handleTagToggle({
-                  name: tag,
-                  isDefault: true,
-                  category: category,
-                })
-              }
-            >
-              {tag}
-            </ToggleGroupItem>
-          ))}
-        </ToggleGroup>
-        <ToggleGroup type="multiple" className="flex flex-wrap gap-3 mt-3">
-          {categoryTagsCustom?.map((tag, index) => (
-            <div key={index} className="flex items-center gap-1">
-              <ToggleGroupItem
-                key={index}
-                value={tag.name}
-                aria-label={tag.name}
-                className={`cursor-pointer p-2 border `}
-                data-state={selectedTag?.name === tag.name ? "on" : ""}
-                onClick={() => handleTagToggle(tag)}
-              >
-                {tag.name}
-              </ToggleGroupItem>
-            </div>
-          ))}
-        </ToggleGroup>
-      </div>
-    ) : (
-      // styling not yet fixed; for now, I´ve applied same styling as for the Tags
-      <p>Keine Tags gefunden</p>
+  const renderTagList = (category) =>
+    renderTagListbyCategoryMetrics(
+      category,
+      standardTags,
+      customTags,
+      selectedTag,
+      selectTag
     );
-  };
 
   return (
     <MetricsContext.Provider
       value={{
-        getAllCheckins,
+        loadCheckinsTotal,
         metricsOneStatus,
         setMetricsOneStatus,
         maxMetricsOneStatus,
@@ -250,7 +167,7 @@ const MetricsProvider = ({ children }) => {
         showMetricsOne,
         setShowMetricsOne,
         previousMetricsOneStep,
-        fetchStatsByTag,
+        loadStatisticsByTag,
         maxMetricsTwoStatus,
         metricsTwoStatus,
         setMetricsTwoStatus,
@@ -263,7 +180,7 @@ const MetricsProvider = ({ children }) => {
         setStatisticsByTag,
         showArrow,
         setShowArrow,
-        renderTagListbyCategoryMetrics,
+        renderTagList,
       }}
     >
       {children}
